requirements.txt
Django==3.2.7
django-filter==2.4.0
django-jazzmin==2.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
